refactor(optimus): Created base class for Search and Select components.
