+++ optimus/lib/optimus.dart
@@ -36,6 +36,8 @@ export 'src/dialogs/dialog.dart';
 export 'src/dialogs/dialog_size.dart';
 export 'src/dialogs/dialog_wrapper.dart';
 export 'src/dialogs/inline_dialog.dart';
+export 'src/dropdown/dropdown.dart';
+export 'src/dropdown/dropdown_tile.dart';
 export 'src/enabled.dart';
 export 'src/expansion/expansion_tile.dart';
 export 'src/form/checkbox_form_field.dart';
@@ -59,7 +61,6 @@ export 'src/number_picker/number_picker.dart';
 export 'src/progress_spinner.dart';
 export 'src/radio/radio.dart';
 export 'src/radio/radio_group.dart';
-export 'src/search/dropdown_tile.dart';
 export 'src/search/search_field.dart';
 export 'src/segmented_control/segmented_control.dart';
 export 'src/select.dart';
