+++ optimus/lib/src/dropdown/dropdown.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:optimus/optimus.dart';
 import 'package:optimus/src/border_radius.dart';
 import 'package:optimus/src/common/anchored_overlay.dart';
+import 'package:optimus/src/dropdown/dropdown_tap_interceptor.dart';
 import 'package:optimus/src/elevation.dart';
-import 'package:optimus/src/search/dropdown_tap_interceptor.dart';
 
 class OptimusDropdown<T> extends StatelessWidget {
   const OptimusDropdown({
